Cambio en tamaños en todos los niveles
Assets, players y Enemigos achicados para hacer que el mapa sea mas grande. Mas espacio para moverse, mejor jugabilidad.

+++ Assets/Scripts/Enemies/Eye scripts/enemyPath.cs
@@ -17,7 +17,7 @@ public class enemyPath : MonoBehaviour
 
 
     [SerializeField] List<Transform> wayPoints;
-    public int Speed = 2;
+    public float Speed = 2;
     public bool NeedShot;
     int nextPos = 0;
